move utils to api project
