--- api/crypto/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-    `java-library`
-    id("org.example.age.java-conventions")
-}
-
-dependencies {
-    api(libs.jackson.annotations)
-    api(libs.jackson.databind)
-
-    testImplementation(libs.guava.testlib)
-}
