+++ api/custom/src/test/java/org/example/age/api/util/AsyncCallsTest.java
@@ -1,9 +1,9 @@
-package org.example.age.service.util;
+package org.example.age.api.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.example.age.testing.CompletionStageTesting.assertIsCompletedWithErrorCode;
 
+import jakarta.ws.rs.WebApplicationException;
 import java.io.IOException;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ExecutionException;
@@ -26,7 +26,9 @@ public final class AsyncCallsTest {
     public void unsuccessfulResponse() {
         Call<Integer> call = Calls.response(Response.error(400, ResponseBody.create("", null)));
         CompletionStage<Integer> asyncValue = AsyncCalls.make(call);
-        assertIsCompletedWithErrorCode(asyncValue, 400);
+        assertThatThrownBy(() -> asyncValue.toCompletableFuture().get())
+                .isInstanceOf(ExecutionException.class)
+                .hasCauseInstanceOf(WebApplicationException.class);
     }
 
     @Test
