+++ service/src/main/java/org/example/age/service/SiteService.java
@@ -19,13 +19,13 @@ import org.example.age.api.VerificationRequest;
 import org.example.age.api.VerificationState;
 import org.example.age.api.VerifiedUser;
 import org.example.age.api.client.AvsApi;
+import org.example.age.api.util.AsyncCalls;
 import org.example.age.service.api.crypto.AgeCertificateVerifier;
 import org.example.age.service.api.crypto.SiteVerifiedUserLocalizer;
 import org.example.age.service.api.request.AccountIdContext;
 import org.example.age.service.api.store.PendingStore;
 import org.example.age.service.api.store.PendingStoreRepository;
 import org.example.age.service.api.store.SiteVerificationStore;
-import org.example.age.service.util.AsyncCalls;
 import retrofit2.Call;
 
 /** Service implementation of {@link SiteApi}. */
